refactor(tasks): improve validation and caching in deleteTaskAction and findById

+++ frontend/src/actions/delete-task-action.ts
@@ -2,22 +2,28 @@
 
 import { drizzleDb } from "@/db/drizzle";
 import { tasksTable } from "@/db/drizzle/schemas";
+import { taskRepository } from "@/repositories/task";
 // import { asyncDelay } from "@/utils/async-delay";
 // import { logColor } from "@/utils/log-color";
 import { eq } from "drizzle-orm";
 import { revalidateTag } from "next/cache";
 
 export async function deleteTaskAction(id: string) {
-  //TODO: Remover log abaixo
+  //TODO: Verificar usuário logado
 
-  //TODO: Aplicar react toastify para mensagens com error
   if (!id || typeof id !== "string") {
     return {
       error: "Dados inválidos",
     };
   }
-  //TODO: Criar validação por ID para saber se a task existe na base de dados (verificar se função findById já existe):
-  // const task = await taskRepository.findById()
+
+  const task = await taskRepository.findById(id).catch(() => undefined);
+
+  if (!task) {
+    return {
+      error: "Tarefa não exista na base de dados",
+    };
+  }
 
   await drizzleDb.delete(tasksTable).where(eq(tasksTable.id, id));
